+++ client/src/GlobalStyles.js
@@ -3,7 +3,9 @@ import { createGlobalStyle } from "styled-components";
 export default createGlobalStyle`
 * {
     box-sizing: border-box;
-    
+    margin:0;
+    padding:0;
+    line-height: 1.5rem;
 }
 
 body {
