+++ client/src/lib/roundPlaces.js
@@ -1,0 +1,4 @@
+export function roundPlaces(float, places) {
+  const power = Math.pow(10, places);
+  return Math.round(float * power) / power;
+}
